Further work upon compute_Rt.py
Tweaked with the calculation so that no errors are thrown as a result of dividing by zero and made it so that any below zero answer is converted to zero.
Implemented a method of handling the Trinary features, where by they are further split into three fields, at which point each of the three levels are handled like binary features.