Some tweaks to work more generally
Adjusted orderFields so that it now isn't hard coded to work with the uk dataset.

Reduced CONFIRMED_THRESHOLD to allow more rows in for lower population areas (before one of the Welsh counties only had one row out of its entire dataset that was present).

Stopped the replacement of the Case number, with the Case Change number. Now place it in its own field.