Corrected the conversion of trinary control measures
Now the values calculated are done in the same way as the binary control measures.

Also, changed it so that when a control measure is increased or decreased, all levels of that control measure below that are still in effect, since the effects of a low control measure are encompassed within a moderate control measure.