Further tweaks to compute_Rt
Changed the calculation of average cases, to be the same as the code it was based on.

Changed how the default value for Rt is given, that some regions contain 0 cases for a few of the first days.

Changed filterDate to use the optDataList due to the filtering not working as intended with newData

Changed getRegionalIndexs to work as intended.