Updated randomForestSHAP
Now it produces the figures for both versions of the dataset at the same time.

It can also save the figures to a png, however, due to the way that tight_layout() works, it results in a less than optimal appearance for the plots.
Without tight_layout(), saving the plots results in the labels being cut off.