Further work on rtPredictor.py
Realised I made a mistake by adding actionListResults[currBestIndex] to the Rt value rather than actionList[currBestIndex].
Which, after correcting, massively improved the accuracy of the result.

Changed the graphs to use only one axis for the Rts, and they now look rather snazzy.

Started working on the next step of the predictor. Which is the evaluation of the impact of each point within a group of a region's data in reaching a goal value.

Currently working on adapting the Greedy Agent's function to help with achieving this goal.