Update rtPredictor.py
Changed greedyAgent:
The index for the action is stored, not the action value itself.

Removed startIndex and endIndex due to a change in how evaluation would be handled.

Implemented initial part for the group evaluation.
Where it produces an alternative attempt at achieving the original Rt value.

Where it iterates and steps forward over the (see GROUP_SIZE) 3 point prior to the goal value. Where at each of the initial points in the iteration, it cannot perform the same action as it did before.
Thus, the current idea is to compare the new attempts at acheiving the goal value, with the original. And then, determine which previous point was most important based upon which one was most affected.