Update rtPredictor.py
Further work upon finishing the evaluation of points within each group of the new Rt values.
Where it finds the values that as a result of the prevention of an action, now are further away in value in comparison to the original agent.

Wrote down an idea for future tree agent implementations, due to the Christmas break.

Merry Christmas and a Happy New Year!
=)