Updated rtPredictor.py
Cleaned up main() by putting the different parts into their own functions.

Prepared parts of the functions for the introduction of new agent types.

Changed how the difference it handled in what is now evaluateAgentPerformance into something that seems to produce results as intended in a more readable manner.

Ran rtPredictor on the newly updated UK dataset, resulting in more days being present for each region and the consolidation of the Welsh counties into Wales.