Worked on tree agent
Further tinkering on the code has lead to vastly improved results.

Realised several mistakes that were present.
Some were due to not making adjustments when the code was changed to iterate over all of a region's data, rather than just an initial one as was the case at the time.
Others just remanants of previous ideas that are no longer be pursued.

Changed the names of the variables used to store the nodes due to not being satified with them