Update rtPredictor.py
Further tweaks:

Changed where the break occurs for last elements within runEvaluationAgent

Changed how the comparison of agent and eval agent difference is done, through the use of abs() in order to get a more desirable result.

Changed the handling of impactful points so that if all points end up being considered impactful, then none are.

Started preparation for the next step of rtPredictor, with writeFile(). In the short term, it will serve as a method of storing and tinkering with the data to help better construct the machine learning and XAI step.