Update rtPredictor.py
Changed how the point impact evaluation is done.
Now rather than only focusing on the most impactful point it orders them within a list.
With the intention being that this list will be used to determine the percentage share of each point's agent difference as the new metric of a points impact.