Update rtPredictor.py
Small tweak to the file saving function, to remove potentially confusing use of "Orig Agent", due to the previous use of "Eval Agent" no longer being present.

Worked on the data preparation for LSTM, where it uses the .csv and gathers it as a Pandas Dataframe.
Then then splits the data, and performs feature scaling to convert all columns (bar the inital No columns and the Binary Point [No] Most Impactful?) into a number between 0 and 1.