Update rtPredictor.py
After wrangeling with Dataframes and trying different ideas:

It now finds the largest Group No that exists in all regions, and cuts off any that go above.

Feature scaling is now properly setup

Started looking into creating the LSTM, but in order to test how well it works with the data, further reshaping needs to be done.