Update compute_Rt.py
After working on rtPredictor.py and looking at the code to get where one region starts and the next ends, I realised that there was a slight error in the code, that resulted in the end index for one region, becoming the start index for another.

A simple -1 was all that was needed to remedy this issue, and make the indexs accurate.

Also, realised an out of index error was bubbling away under the surface but had yet (until now) arisen due to how the indexes were used before.