Update rtPredictor.py
Further work upon the preprocessing for the LSTM.

Where I have done a large chunk of the work towards having the data put together all of the group data into their own sublist.

The current setup, while it can be run with the LSTM, it produces unusable results and was explored out of curiousity rather than expecting it to be functional.
Where it splits the data within each sublist, rather than by the groups.
Meaning that within each group currently there are only 2 points rather than 3, with the 3rd element being put in the test data.