Update rtPredictor.py
Further tweaking of the LSTM

Made a method of seeing the accuracy based upon how correct its predictions were on a per group basis (whether it was or wasn't an impactful point).
And, made it so that the output of that function adjusts to differing sizes of data and outputs.