Update rtPredictor.py
Validation data is now created from half of what would have originally been the Test data.
In regards to the LSTM:
After lots of testing and tweaking of the LSTM, I have come up with a solution that has improved the accuracy by around 5%.

Standard activation function is used for the input layer, but Swish is used for the second LSTM layer over that of ReLU.

Changed from standard optimizer, to using optimizer object that allows further tweaking.

Upped batch_size to 12, due to it providing increased accuracy, along with less turbulent and more promising loss.

Now the loss is plotted from each run of the model, using the history given by model.fit().

Also, for the sake of mass testing regions, I implemented two for loops that iterate and average the accuracies of the test data, then print this information at the end (currently, it is setup to run all of the data, rather than each region, to change this simply change regionEndNo to 13 to run all of them, and put currRegion at the end of prepareData()