Update rtPredictor.py
Further tweaks to the LSTM that have resulted in more promising and consistent results.

Doubled the number of neurons in the Dense layer.

Quadrupled the number of neurons in the LSTM layers.

Halved the dropout for the input layer and Dense layer.

Increased the size of the test chunk,  and reduced the share of the validation chunk.