Changed goal from prediction impact point, to predicting next action
In order to faciliate this, it has required a little re-work of several parts through out rtPredictor.py

The changes are as follows:

The consolidation of a groups points back into one row, along with the data being saved into each row not pertaining to a point impact.
Which is replaced with information about the point after the goal point, which is the action that needs to be predicted.

Due to the change in goal being only Yes or No, to a potential choice of QUANTILE_NO *2 + 1, it has required a rework of parts relating to the handling of the data.

Finally, due to the focus now being on the actions themselves, for the sake of simplifying things, I have changed QUANTILE_NO from 8 to 5, resulting in a change of potential actions from 17 to 11.