Update compute_Rt.py
Fixed Trinary bugs and oversights.

Now the Trinary values are being calculated as intended and the use of date reduction has been brought in line with what is intended. Where now it will no longer keep on checking further and further back if a previous date reduction for a control measure (or level of that control measure) did not satisfy the necessary conditions.