Update rtPredictor.py
Altered how the Rt agent results are saved, so that each row contains WINDOW_SIZE actions, rather than being determined by the group sizes.

Created a new function that is specifically for the testing of different hyper parameters.

Created a function to create Sine waves in order to test both the agents and the LSTMs based upon the agents chosen actions to mimic the Sine wave.

Created a custom loss function RMSE which will be used going forward to see if it provides an increase to accuracy.

Moved the creation of graphs for loss to outside of the testLSTM function for the sake of consolidating Training and Validation into their own graphs.