Further exploration of balancing techniques
Focused more on looking at Oversampling techniques to get a roughly equal amount to the under sampling techniques.

Cleaned up prepareDataRegressor (previously prepareDataRandomForest) due to it becoming rather messy over the course of testing new techniques.

Added in splitting the core data into two parts, so that the training data can have the balancing techniques run on them and there remains a portion of the dataset that is guaranteed to have not been seen by the the model.

Made a function for the BaggingRegressor so it can be called directly from the Notebook.

Deleted window sizes that are no longer needed, and are currently just cluttering up the folder.