Changed how rtPredictor outputs images and changed file structure
Made it so that now different datasets will now output the images to their own sub-folder, rather than storing and in turn overwriting the images in one folder.

Changed the folder structure of the images to something more sensical, where the only reason for not doing so earlier was to prevent the historical loss of data for those files.
Which during development made comparing the current image to the one stored in the repository a useful tool.

Also reduced the size of the variables for prepareDataRegressor, due to their names being unnecessarily long.