Tweaks to rtPredictor.py and further Testing of Datasets
Testing of the Financial Dataset has lead to further adjustments to the structure and hyperparameters of the LSTM and the data pre-processing.

Fully introduced the Sine Data as a proper dataset, rather than just a small function due to it being a useful thing to keep a standard location for results.

This should also be my final proper commit (though knowing how it went at Christmas, there might be something else) to the repository.
So to whoever reads this, hope you're doing well and good luck in what you do!